pyproject.toml
[tool.poetry]
name = "webtest"
version = "0.1.0"
description = ""
authors = ["Mariana Dias"]

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
selenium = "^3.141.0"
robotframework = "^4.1"
robotframework-selenium2library = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
